DCS-1002 Derive court name from court id where possible.
