+++ src/main/resources/db/migration/V23__revise_court_ids.sql
@@ -1,0 +1,11 @@
+delete from enabled_court where id = 'DRBYMC';
+
+update enabled_court set id = 'DRBYMC' where id = 'DRBYJC';
+
+update video_link_appointment vla
+    set court_id = (
+            select ec.id
+            from enabled_court ec
+            where ec.name = vla.court
+        )
+where vla.court_id is null;
